requirements.txt
# Core ML
numpy==1.26.4 --only-binary=:all: --index-url https://pypi.org/simple


pandas==2.1.4
scikit-learn==1.8.0
python-dotenv==1.0.1
httpx==0.27.0
pytest-asyncio==0.23.6


# API
fastapi==0.110.0
uvicorn==0.27.1

# Experiment tracking
mlflow==2.10.2

# Visualization
matplotlib==3.8.3
seaborn==0.13.2

# Testing
pytest==8.1.1

# Serialization
joblib==1.3.2

# Monitoring (optional but allowed)
prometheus-client==0.20.0


